pyproject.toml
[project]
name = "research"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.3",
    "openai>=2.17.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlite-vec>=0.1.6",
    "uvicorn[standard]>=0.40.0",
]
